requirements.txt
aiogram
asyncpg
requests
pandas
matplotlib
python-dotenv
xlrd
openpyxl
